♻️ UserDto 리팩토링 (#2)
