--- src/main/java/com/example/carefully/domain/user/dto/RoleRequest.java
@@ -1,10 +1,0 @@
-package com.example.carefully.domain.user.dto;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum RoleRequest {
-    USER, OPERATION
-}
